requirements.txt
django==3.1.2
pillow==7.2.0
django-admin-rangefilter==0.6.3